move to legacy: node_ref
