+++ weave/legacy/node_ref.py
@@ -3,7 +3,7 @@ import typing
 
 from weave.legacy import graph
 
-from . import ref_base
+from .. import ref_base
 
 # Notes for the future:
 # - I added list.lookup to lookup rows in a list by ID. I think we probably should
